+++ GgafDx9Core/src/jp/ggaf/dx9core/util/GgafDx9Quaternion.cpp
@@ -3,7 +3,7 @@ using namespace std;
 using namespace GgafCore;
 using namespace GgafDx9Core;
 
-GgafDx9Quaternion::GgafDx9Quaternion(double prm_t, double prm_x, double prm_y, double prm_z) : GgafObject(),
+GgafDx9Quaternion::GgafDx9Quaternion(float prm_t, float prm_x, float prm_y, float prm_z) : GgafObject(),
     _t(prm_t),
     _x(prm_x),
     _y(prm_y),
